requirements.txt
python-binance==1.0.16
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0
python-telegram-bot==13.15
dash==2.9.3
plotly==5.14.0
matplotlib==3.7.1
